+++ modules/registration_module/registration_functions.py
@@ -124,11 +124,17 @@ def registration_with_tmat(tmat_int, image, skip_crop, output_path):
         y_end = image.sizes['Y'] - max([d[2] for d in tmat_int if d[3] == 1])
         x_start = 0 - min([d[1] for d in tmat_int if d[3] == 1]) 
         x_end = image.sizes['X'] - max([d[1] for d in tmat_int if d[3] == 1])
-        t_start = min([d[0] for d in tmat_int if d[3] == 1])
+        t_start = min([d[0] for d in tmat_int if d[3] == 1]) - 1
         t_end = max([d[0] for d in tmat_int if d[3] == 1])
 
         # Crop along the y-axis
         image_cropped = registered_image[:, t_start:t_end, :, :, y_start:y_end, x_start:x_end]
+        print(image6D.shape)
+        print(registered_image.shape)
+        print(image_cropped.shape)
+        print(t_start,t_end)
+        print(y_start,y_end)
+        print(x_start,x_end)
 
         # Save the registered and cropped image
         OmeTiffWriter.save(image_cropped[0,:,:,:,:,:], registeredFilepath, dim_order="TCZYX")